refactor: 컴포넌트 및 아이콘 경로 리팩토링
import 대신 src에 public 이하 경로로 접근할 수 있도록 수정해주세요.
