+++ frontend/public/icons/category.png
